python/requirements.txt
moviepy==1.0.3
librosa>=0.11.0
numpy>=2.3.3
pillow>=11.3.0
ffmpeg>=1.4
flask
flask-cors
requests